token added to the blog part

+++ functions/delete_post.php
@@ -4,12 +4,14 @@ include_once ''.$_SERVER['DOCUMENT_ROOT'].'/SecureBlog/sql/blogMapper.php';
 
 sec_session_start();
 
-if(isset($_GET['id']))
+if(isset($_GET['id']) && isset($_GET['token']))
 {
     $post_id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT);
     $user_id = $_SESSION['userID'];
     
-    if(i_am_owner_of_post($mysqli, $post_id, $user_id))
+    $token = $_GET['token'];
+    
+    if($token == $_SESSION['token'] && i_am_owner_of_post($mysqli, $post_id, $user_id))
     {
         if(deletePost($mysqli, $post_id))
         {
